feat: use typescript for react-app-rewired config
