--- headers/index.js
@@ -1,8 +1,0 @@
-const tsNode = require('ts-node')
-tsNode.register({
-  compilerOptions: {
-    module: 'CommonJS',
-  },
-})
-
-module.exports = require('./index.ts')
