+++ package.json
@@ -103,7 +103,7 @@
     "show-linked-packages": "find node_modules node_modules/@* -maxdepth 1 -type l -print",
     "link-packages": "./scripts/link-packages.sh",
     "unlink-packages": "find node_modules node_modules/@* -maxdepth 1 -type l -print | awk -F/ '{ printf \"%s/%s\\n\",$2,$3; }' | xargs yarn unlink && yarn install --force",
-    "build": "NODE_OPTIONS=--max-old-space-size=4096 IMAGE_INLINE_SIZE_LIMIT=0 INLINE_RUNTIME_CHUNK=false DISABLE_ESLINT_PLUGIN=true react-app-rewired build && node ./headers/write.js && ./scripts/sha256sums.sh && ./scripts/verifyWebpackHashes.sh",
+    "build": "NODE_OPTIONS=--max-old-space-size=4096 IMAGE_INLINE_SIZE_LIMIT=0 INLINE_RUNTIME_CHUNK=false DISABLE_ESLINT_PLUGIN=true react-app-rewired build && node ./scripts/writeHeaders.js && ./scripts/sha256sums.sh && ./scripts/verifyWebpackHashes.sh",
     "dev": "NODE_OPTIONS=--max-old-space-size=4096 ESLINT_NO_DEV_ERRORS=true TSC_COMPILE_ON_ERROR=true IMAGE_INLINE_SIZE_LIMIT=0 INLINE_RUNTIME_CHUNK=false CYPRESS_RELAXED_SECURITY=true react-app-rewired start",
     "dev:silent": "BROWSER=none yarn dev",
     "local-ci": "yarn lint && yarn type-check && yarn test && yarn build",
@@ -112,7 +112,7 @@
     "lint:fix": "yarn lint --fix",
     "lint:cypress:fix": "yarn lint:cypress --fix",
     "minify": "yarn yarn-minify",
-    "test": "yarn jest --preset ts-jest headers && react-app-rewired test --coverage --watchAll=false",
+    "test": "yarn jest --preset ts-jest react-app-rewired && react-app-rewired test --coverage --watchAll=false",
     "test:cypress": "yarn cypress open",
     "test:cypress:headless": "yarn cypress run --headless",
     "test:cypress:rerecord": "yarn cypress run --browser chrome --headless --env forceRecord=true",
@@ -180,6 +180,7 @@
     "@testing-library/react-hooks": "^7.0.1",
     "@testing-library/user-event": "^13.2.1",
     "@types/bignumber.js": "^5.0.0",
+    "@types/circular-dependency-plugin": "^5.0.5",
     "@types/crypto-js": "^4.0.2",
     "@types/d3-array": "^3.0.1",
     "@types/d3-time-format": "^4.0.0",
@@ -201,6 +202,7 @@
     "@types/react-virtualized-auto-sizer": "^1.0.1",
     "@types/react-window": "^1.8.5",
     "@types/redux-logger": "^3.0.9",
+    "@types/ssri": "^7.1.1",
     "circular-dependency-plugin": "^5.2.2",
     "cypress": "10.0.2",
     "eslint-config-prettier": "^8.3.0",
